basic building block for two different document views
