+++ src/components/Documents/SearchQueryForm.vue
@@ -1,0 +1,6 @@
+<template>
+  <q-card-section class="row q-pa-none" />
+</template>
+
+<script lang="ts">
+</script>
