+++ src/i18n/de.json
@@ -316,7 +316,14 @@
     "emptyResult": "Kein Ergebnis|Keine ergebnisse"
   },
   "repositoryType": "Repositoriumstyp|Repositoriumstypen",
-  "documentSearch": "Dokumentensuche",
+  "documentSearch": {
+    "title": "Dokumentensuche",
+    "description": "Bitte wählen sie die Art der Dokumentrepräsentation aus.",
+    "type": {
+      "conceptCluster": "Konzept-Cluster",
+      "searchQuery": "Such-Abfrage"
+    }
+  },
   "concept_repository": "Konzept Repositorium",
   "phenotype_repository": "Phänotyp Repositorium",
   "exclusive": "Ausschließlich",
